Moving theme code into a manager class
